requirements.txt
pillow
streamlit
mira-sdk
python-dotenv
